Re-organize query params and query extensions
