+++ LBH.AdultSocialCare.Api/V1/Infrastructure/RequestFeatures/Parameters/BrokerPackageViewQueryParameters.cs
@@ -1,7 +1,7 @@
 using LBH.AdultSocialCare.Api.V1.AppConstants.Enums;
 using System;
 
-namespace LBH.AdultSocialCare.Api.V1.Infrastructure.RequestExtensions
+namespace LBH.AdultSocialCare.Api.V1.Infrastructure.RequestFeatures.Parameters
 {
     public class BrokerPackageViewQueryParameters : RequestParameters
     {
