+++ LBH.AdultSocialCare.Api/V1/Infrastructure/RequestFeatures/Parameters/RequestParameters.cs
@@ -1,4 +1,4 @@
-namespace LBH.AdultSocialCare.Api.V1.Infrastructure.RequestExtensions
+namespace LBH.AdultSocialCare.Api.V1.Infrastructure.RequestFeatures.Parameters
 {
     public class RequestParameters
     {
