+++ LBH.AdultSocialCare.Api/V1/UseCase/Clients/Interfaces/IGetAllClientsUseCase.cs
@@ -1,6 +1,6 @@
-using System.Threading.Tasks;
 using LBH.AdultSocialCare.Api.V1.Boundary.Common.Response;
-using LBH.AdultSocialCare.Api.V1.Infrastructure.RequestExtensions;
+using LBH.AdultSocialCare.Api.V1.Infrastructure.RequestFeatures.Parameters;
+using System.Threading.Tasks;
 
 namespace LBH.AdultSocialCare.Api.V1.UseCase.Clients.Interfaces
 {
